requirements.txt
aircv==1.4.6
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
ddddocr==1.4.3
Deprecated==1.2.13
execnet==1.9.0
flatbuffers==2.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
numpy==1.21.6
onnxruntime==1.11.1
opencv-python==4.1.0.25
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
protobuf==4.21.5
py==1.11.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.2
pytest-assume==2.4.3
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==2.0.2
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-dotenv==0.20.0
PyYAML==6.0
redis==4.3.4
requests==2.28.1
selenium==3.141.0
six==1.16.0
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
webdriver-manager==3.8.3
wrapt==1.14.1
zipp==3.8.1
